backend/ai_service/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
numpy
rembg
transformers
omegaconf
einops
scikit-image
